+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 * Added `segmentedCollectionView(_:willDisplayHeader:)` to `SegmentedCollectionViewDelegate`
 * Improved documentation of `SegmentedCollectionViewDelegate`
+* Added `FlaneurSearchViewController` and the associated demo
 
 ## [0.8.0] - 2018-02-05